Update author information in project metadata and package init
Changed the author name and email in pyproject.toml to "Arthur Souza Rodrigues" and "arthrod@umich.edu" respectively. Updated the __init__.py file in the newdotfiles package to reflect the same author name and email, replacing the placeholder values. Also added a proper module docstring describing the project as a Python boilerplate project.